# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Retail CRM API",
        "description": "API для управления розничной CRM-системой: товары, склады, заказы, поставки, перемещения, клиенты.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base path"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Список заказов",
                "description": "Возвращает заказы с пагинацией и фильтрацией по статусу, клиенту, складу, дате создания.",
                "operationId": "8f3183bfa0ecb8a2c864cbf972b8061d",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "completed",
                                "canceled"
                            ]
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Фильтр по клиенту",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Фильтр по складу",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "created_from",
                        "in": "query",
                        "description": "Дата создания от (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "created_to",
                        "in": "query",
                        "description": "Дата создания до (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Кол-во на странице (по умолчанию 15)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пагинированный список заказов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Создать заказ",
                "description": "Создаёт заказ со списком позиций. Проверяет наличие товара на складе. При недостатке возвращает 422 с unavailable_products.",
                "operationId": "b6490f9d9026dbe871b41207e118fffb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заказ создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации / недостаточно товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnavailableProducts"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Детали заказа",
                "description": "Возвращает заказ с позициями, клиентом и складом.",
                "operationId": "79c1e32fc3603170dfd7659562133aeb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Обновить заказ",
                "description": "Обновляет заказ (только со статусом active). При обновлении позиций — пересчитывает движения остатков.",
                "operationId": "be1c75741bfcbee84e078c24a476f6ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заказ обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Заказ не в статусе active"
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}/complete": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Завершить заказ",
                "description": "Переводит заказ в статус completed. Списание остатков и запись движений выполняются при завершении.",
                "operationId": "d17bc7166cd113ab0cef698d34b88245",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ завершён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Заказ не в статусе active",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Only active orders can be completed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}/cancel": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Отменить заказ",
                "description": "Переводит заказ в статус canceled.",
                "operationId": "37236986b645db71797c4af565697a26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ отменён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Заказ не в статусе active",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Only active orders can be canceled."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}/resume": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Возобновить заказ",
                "description": "Возвращает отменённый заказ в статус active. Проверяет наличие товара на складе.",
                "operationId": "2299b7e63866e0fb4f51f7401114eeab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ возобновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Заказ не в статусе canceled / недостаточно товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnavailableProducts"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Футболка белая"
                    },
                    "count": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "warehouse_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "warehouse_name": {
                        "type": "string",
                        "example": "Основной склад"
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "completed",
                            "canceled"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "datetime",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            },
...
            "StoreOrderItem": {
                "required": [
                    "product_id",
                    "count"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "count": {
                        "type": "integer",
                        "example": 3,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
            "StoreOrderRequest": {
                "required": [
                    "customer_id",
                    "warehouse_id",
                    "items"
                ],
                "properties": {
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "warehouse_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreOrderItem"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateOrderRequest": {
                "properties": {
                    "customer_id": {
                        "type": "integer"
                    },
                    "warehouse_id": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreOrderItem"
                        }
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UnavailableProducts": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Some products are not available in sufficient quantity."
                    },
                    "unavailable_products": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "product_id": {
                                    "type": "integer"
                                },
                                "required": {
                                    "type": "integer"
                                },
                                "available": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}